+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Notification of other windows.
 - Rewards.
 - Triggers.
+- Added methods for obtaining purchases and achievements.
 
 # [v1.1.2](https://github.com/megalanthus/defold-gamepush/releases/tag/v1.1.2) (2023-04-05)